os: change Input::Touch_id from int to unsigned
There no sensible meaning for negative touch IDs.

Issue #3669

+++ repos/dde_linux/src/drivers/usb_hid/main.cc
@@ -281,7 +281,7 @@ void Driver::input_callback(Input_event type, unsigned code,
 	case EVENT_TYPE_WHEEL:   submit(Wheel{rx, ry});           break;
 	case EVENT_TYPE_TOUCH:
 		{
-			Touch_id const id { (int)code };
+			Touch_id const id { code };
 
 			if (rx == -1 && ry == -1)
 				submit(Touch_release{id});
